Zero particle orphans (#23)
* Add back tracer if all removed during masking

* Update tracerIndex

* Set formatter version

* Update Tracer position and index for central subhaloes

* Initialize TracerIndex to -1 to catch use before setting

* Add back most bound remaining tracer after masking

* Save tracer index before swapping particles

* Update modules for testing

* Formatter

* Make it so logs get stored in the relevant run directory

* Clarify masking only occurs for pre-existing subhalos

* run_test.sh now takes (optional) parameter
It specifies where to save results.

* Fixed variable expansion in sed call.

* Use tracer index when masking satellites

* Iterate over all particles to find most bound tracer

* Find most bound tracer in satellites

* TracerIndex is correct

* Using multiple collisionless particles to decide host.
Currently OMP-only implementation.

* Changed the workflow to identify particle IDs
This will allow us to generalise the way in which is done
to MPI runs.

* Fixed default value of MinNumTracerPartOfSub.
It was 20, when it should have been 10.

* TracerID vector is now created with correct size.
Before, it was being resized within the function to assign ParticleIds

* Added comments

* Started MPI implementation
So far, we obtain TracerIds of local subhaloes, to be communicated later

* Communicate TracerIDs.
Currently, need to turn 2d vector to 1d vector before MPI broadcast. In future, change vector
dimensions to be 1d from the get-go, or see if MPI can handle 2D cases easily
(apparently not)

* Changing functions to enable future 1D vector usage.

* Split functions into several ones, to enable MPI implementation
Essentially, we will now be able to use a 1d vector from the get-go. This will
circumvent the need of flattening a 2D vector before communicating

* Changed ParticleID vector from 2D to 1D.
This will prevent having to flatten it at a later time

* Function to assign TracerParticleIds now works with iterators

* Changed GetTracerHosts to work with iterators

* Made DecideLocalHostId work with iterators.
OMP-only implementation now working with iterators.

* MPI populates 1D ParticleID vector, and broadcasts it.
No more need to flatten a 2D vector.

* Forgot to commit dtype change in function variable

* Can now decide Host and Rank in MPI mode.

* Formatting

* Added sanity checks when getting Tracer ParticleIds

* Number of tracers used for host finding can be specified in config.
If unspecified, it will default to the minimum number of tracers required in each subhalo.

* Added postprocessing test to check internal HostHaloId decisions.

* Added test to identify tracks with shared particles.
It also returns the fraction of shared particles

* Removed personal imports

* MPI warning updated to reflect extra data being sent.

* Disable host decisions if we have a -1 or -2 in local search
Previously, it would only disable it if we have -2

* Orphans are set to zero particles.
This happens if they are found to be unbound during unbinding.

* Removed unneeded function call
If subhaloes are found to have less particles than the threshold,
it is dealt with at a later stage.

* Re-arranged code
We inmediately skip Unbind if we have an orphan.

* Handle the Host identification of orphans

* Added claryfing note
I got confused about what to do about fake subhaloes

* Ensure orphan coordinate is at previous tracer
I think previous version could've returned the position of
the most bound particle (!= position of previous most bound tracer)

* Create zero-particle orphans during merging

* Changed indexing to set MostBoundParticleId
It does the same as before, but uses GetTracerIndex instead.

* Changed sanity checks in GetTracerIds
Oherwise it will trigger for zero-particle orphans

* Run host identification for all subgroups
Before, it would skip zero-particle orphans.

* Tracer identification in FeedCentral works for orphans
It would skip orphans since they now have no associated particles.

* Make testing script compile HBT in Debug mode

* Assertion now executes for resolved subhaloes.
It will trigger an error for zero-particle orphans, as expected.

* Orphans get correct position and velocity for merging
Before, they would be skipped.

* Set the Average position and velocity of orphans

* Added comments

* Set the initial most bound position of resolved subhaloes
Based on the most bound tracer, rather than the most particle.

* Added clarifying comment

* Do not update position of new orphan.
Use the information from when particles were initially updated.

* Set mass of potential orphan during particle update.

* Do the MostBoundParticleId assignment in Unbind
For bound subhaloes it would be done before saving.

* Do not reassign the MostBoundParticleID if we have a disruption

* Added comments

* Fixed bug that would overwrite MostBoundParticleId

* Updated host halo identification test script
It would crash due to the addition of zero-particle orphans. Solved by skipping them

* Added sanity check and comment

* Update configuration file for test

* Fixed method to copy orphan particle velocities
HBT would not compile if HBT_REAL8 and HBT_REAL4_VEL were both enabled

* Applied formatter

* Created MPI tests for host halo finding and consistent MostBoundParticleIDs

* Removed deprecated testing script.
Prefer to use the MPI-enabled one from the previosu commit

* Removed unused functions in testing of orphan ID consistency

* Simpler handling of orphan creation of merged objects.
We now do a single resize, rather than a swap and two resizes.

---------

Co-authored-by: robjmcgibbon <robjmcgibbon@gmail.com>
Co-authored-by: John Helly <j.c.helly@durham.ac.uk>