Remove duplicate particles in mergers (#32)
* Comments to explain recursive unbinding function

* Added pseudocode

* Added function arguments and return values

* Added recursive loop

* Added logic to handle merger testing and property calculation.

* Refactored new merging code.

* Added missing flag

* Modified variable name to more make self-explanatory

* Declared Subhalo_t variables used in new merging method

* Added function definition to measure core phase space properties

* Added function calls to compute phase space properties of the core

* More refactoring

* Do not check phase-space overlap between an subhalo and itself
Duh

* Added missing PCB correction
Would have affected absolute position values, and potentially the position dispersion (if the core of a subhalo is split across a PCB)

* Added (temporary) tests of phase space distance calculations
To check values relative to old calculations. Run disabling the new
merging scheme, to obtain the same subhalo-subhalo comparisons.

* Remove temporary asserts

* Helper function to identify TrackIds across different HBT catalogues
Should only be used for runs using the same FOF and particle information!

* New merging method does not need to globalise SinkTrackId
Information available within Unbind, whereas previous method relied on a
helper array that did not have the same ordering.

* Do not consider unbound subhaloes as eligible to accrete particles through mergers
This is the same behaviour implemented in HBT before.

* Remove unnecessary functions
All removals concern previous merging method.

* Update unit test for phase-space property calculation

* Comments

* Changed main function name

* Removed done TODO

* Missing ")" from comment

* Applied formatter

* Lambda criterion for removing foreing FOF particles defined separately
Change made to help debugging.

* Fixed bug involving hostless subgroups when constraining them to a FOF
This would occur because we tried to access element  -1 in the halo array.

* Added assert to check if we have found all required IDs for merger_tree

* Re-order where single FOF constrains are applied.
Doing so before storing ids in merger tree can lead to some tracers being missing

* Pass bound particles instead of bound + unbound in mergers

* New masking step to remove duplicate particles (#35)

* Masking function that gives preference to shallower subhaloes in the hierarchy

* Added masking from bottom to top

* Update value of TracerIndex to reflect change in Particle vector ordering

* Added function call to mask particles from the source subhalo
Called after saving bound properties.

* Skip FOF groups that only have a central or an unbound component

* Comments

* Moved location of CleanTracks
Makes more sense to be placed within save routine.

* Fixed array indexing for hierarchical relationships
At the stage where we want to do the mask cleaning, the NestedSubhalos values have been globalised and thus do not reflect the location in the local Subhalo array.

* tracer_counter only asserted for subhalos with Nbound > 0

* Update Nbound value after mask clean step.

Outdated value would lead to memory issues in Subhalo_t::KickNullParticles,
which relies of Nbound instead of Particles.size(). We should really make Nbound
consistent throughout the code!!

* Nasty bug with iterator start
Shout out to Rob for spotting it!

* Value of Nbound now means the number of remaining bound particles in the masked subset.

* VERY crude way to allocate sufficient memory in ExclusionList

* Split IO into two distinct steps: bound and source particle saving
This allows me to clean the particles after saving bound properties, but
before saving the source information used for restarts

* Removed stray timer

* More comments

* Fix in duplicate particle testing script.
Would only account for one of the TrackIds hosting a duplicate particle.

* Applied formatter

* Number of tracers used for merger tree default to minimum number of tracers
This value defaults to 10.

* Made the masking work.
- New debug checks added
- Masking from top to bottom and viceversa are called separately, which
  solves cases where the initial central guess was incorrect.
- Fixed a bug when checking whether the number of bound particles is
  changing when masking

* Formatting

* Avoid using MemberTable in CleanTracks()

* Move CleanTracks() call to just before AssignHosts

This avoids problems with restarting because we don't write an out
of date Nbound to the SubSnap files. It also means that we use all
of the bound particles from computing halo positions and velocities
for deciding centrals.

* Fix concerning orphan masses in central identification

Previously, newly created orphans would retain their bound masses,
meaning that they could remain as centrals even if another resolved
subhalo was present. This fix makes all orphans have zero mass, in
line with their zero bound particles.

* Updated comments

---------

Co-authored-by: John Helly <j.c.helly@durham.ac.uk>

* Remove special treatment of orphans in RecursiveUnbind

* Removed commented out code

* Changed default method to insert particles in MergeTo

* Formatter

* Missing call to starting Particle iterator

* Fixed double call of MergeTo when running in DEBUG

It would produce duplicate particles since we'd be adding the bound
component twice. Then the assert in the masking would fail due to the
presence of duplicates and the fail to insert some particles.

---------

Co-authored-by: John Helly <j.c.helly@durham.ac.uk>