Modify MaxConcurrentIO implementation for swift
The previous implementation split the MPI ranks into groups and
prevented any ranks in the next group from proceeding until all
ranks in the current group have finished. This will waste a lot
of time if one (or a few) ranks are very slow.

This new implementation tries to make sure that we always have
MaxConcurrentIO ranks reading. When any one rank finishes the
next is allowed to start immediately.