Handle splitting particles in swift (#47)
* Added script to retrieve and save splitting information

The output is intended to be loaded by HBT, which will then create a map
to include the split particles in the relevant subgroup.

* Removed leftover prints and incorrect for loop range.

* Script for splitting info now loads the HBT config

* Function to get split file name

* Function to open split information fil

* Split file provides number of splits.

Also, the suffix of files now matches snapshot number,
rather than HBT snapshot index.

* Split informatio file saves keys and values in separate datasets

This is easier to load within HBT

* Function now loads splitting information

* Added parameter to enable/disable particle splitting

This can be used to disable particle splitting in swift hydro runs. The
value will be irrelevant for any other snapshot formator DMO simulation

* Load the split information if HYDRO swift simulation and parameter is enabled.

* Pass unordered map to splitting load function

* Subhalo Particles now include particles that have been split

* Removed snapshot format conditions from update particles.

* Create unordered_map to hold ids of split particles

* Changed the way in which swiftsim defaults to particle splitting

* Added more informative error message.

* Formatting

* Added readme for swiftsim toolbox.

* Docs and formatting for python script

* Fixed typo in the README of particle splitting.

* Parameter controlling whether particles split is always saved.

* Final (?) update to how the splitting parameter gets set

Previous version would always make HBT+ use splitting for swift
simulations. The new commit makes the code check whether the parameter
has been specified, and if not, defaults to true in swift.

* Change write mode of hdf5 files in particle split script.